Commit_Hana

+++ pidev/src/main/java/tn/eprit/spring/services/ILeaveTestService.java
@@ -6,11 +6,11 @@ import tn.esprit.spring.entity.LeaveTest;
 
 public interface ILeaveTestService {
 
-	int addLeave(LeaveTest leaveTest);
+		Long addLeave(LeaveTest leaveTest);
 
 	List<LeaveTest> getAllLeaves();
 
-	void deleteByLeaveId(int leave_id);
+	void deleteByLeaveId(Long leave_id);
 
 	LeaveTest updateLeave(LeaveTest L);
